pyproject.toml
[project]
name = "cleopy"
description = "Python package for generating input and processing output of CLEO SDM"
dynamic = ["version"]
authors = [
	{name="Clara Bayley", email="yoctoyotta1024@yahoo.com"},
]
readme = "README.md"
repository = "https://github.com/yoctoyotta1024/CLEO"
license = "BSD-3-Clause"
license-files = ["LICENSE.md"]
requires-python=">=3.13"

dependencies = [
  "awkward",
	"numpy",
  "matplotlib",
  "pathlib",
  "ruamel.yaml",
	"scipy",
  "typing",
  "xarray",
  "zarr",
]

[project.optional-dependencies]
examples = [
  "argparse",
  "mpi4py",
]
yac = [
  "cython",
  "datetime",
  "isodate",
  "netcdf4",
  "mpi4py",
]

[dependency-groups]
dev = [
  {include-group = "pre-commit"},
  {include-group = "test"},
  {include-group = "docs"},
]
pre-commit = [
    "pre-commit",
]
test = [
    "pytest",
]
docs = [
    "breathe",
    "furo",
    "sphinx",
    "sphinx_copybutton",
    "sphinxcontrib-applehelp",
    "sphinxcontrib-bibtex",
    "sphinxcontrib-htmlhelp",
]

[project.urls]
Homepage = "https://github.com/yoctoyotta1024/CLEO"
Website = "https://yoctoyotta1024.github.io/CLEO"
Issues = "https://github.com/yoctoyotta1024/CLEO/issues"

[tool.pytest.ini_options]
pythonpath=[
  ".", "cleopy"
]

[build-system]
requires=["hatchling >= 1.26", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"
